Remove intermediate docs
